pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flyer-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"numpy>=2.0",
	"opencv-python>=4.10",
	"pillow>=10.4",
	"pyyaml>=6.0",
	"weasyprint>=62.0",
]

[project.scripts]
flyte = "flyte.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
